Address memory exception around validate error
ValidateAndCompareConfiglets doesn't have strict type on return fields.
Successful case yields a string for errors, however in failure cases the
field is an array of strings. This is seen for other fields also. Fix is
to handle the error returns and ignore the other fields since this
appears to be the behavior of the API.